refactor(header-menu): update file paths
